change option parsing to clap
Prior to this commit we were using getopts for option parsing, which,
while it made it easy to detect options passed to the application,
getopts didn't have any functionality for stating that certain arguments
were required or the cardinality of the those arguments. This meant that
we'd have to manually traverse the command line arguments and do that
validation ourselves. This commit replaces the usage of getopt with clap
in order to delegate that work of argument validation and cardinality.