Add RSL init and push methods
RSL init just creates an (almost) empty branch. Since the git2rust library
requires a valid commit reference to be passed to the branch method, we will
have to find a workaround or get the library updated in order to create a root
commit. For now, the branches are created off the initial repository commit.
Also, refactor 'fetch rsl and nonce from remote' method.