Implement Serde serialize an deserialize for PushEntry
The macros for implementing Serialize and Deserialize on the PushEntry struct
require a like implemenation for all types used in the struct. Since the struct
contains git2 types for which ser/deser are not implemented, they had to be
handled using this workaround: https://serde.rs/remote-derive.html. This commit
uses the Ser and Deser implementations to parse PushEntries from git objects, strings,
and back again.

In order to do so initially, some of the types inside the pushentry struct
were changed. `head` was changed to always be an Oid, and `related_commits`
was removed altogether, as these are complex types containing foreign types for
which `serialize` and `deserialize` had to be implemented using the workaround.
If possible, these should be added back in later. Additionally, the `nonce_bag`
field now uses the NonceBag struct instead of a HashSet.