Add serialization tests
Prior to this commit, the tests for to and from string for PushEntry sometimes
failed because of the unpredicatable order of serializing a HashSet. This
commit updates the tests to use a different strategy, first serialzing then
deserializing and testing the output. It also updates the test data to reflect
the data in new git fictures.