Finally rid of al compilation errors, only to arrive at the real problem
The real problem is the lack of theough and intentional design in ownership of
RSL and associated structs and objects throughout their lifetime during the run
of a single command. The common library has been mostly refactored into RSL
traits on the repository type. The next step is to update the RSL type to include
a nonce bag, a remote rsl, and a local rsl. This will make passing objects to
routines much easier, and clean up the external interfaces by only providing
capabilities to the RSL, and nothing else.