Fix implementation of commit
Prior to this commit, the method that commited a oush entry to the local
RSL for a repository always returned an error, since it was unable to find the
object pointed to by the provided parent commit reference. This commit fixes
how the parent commit is determined. It also changes the test setup to return
an unwrapped Repository. However, the implementation still isn't finished, as
it appears the head of the RSL branch is not updated upon commit.