Stash untracked files
Prior to this commit, the prep workspace routine returned an error if there
was nothing in the index to stash. This commit ensures that a stash is only
attempted if there are changes in the index or working tree. It also is
improved to stash untracked files in the work tree.