Distinguish between global and local init
Prior to this commit, the only init case considered was when nobody had ever
created an RSL for the repo before. This commit adds a case for RSL existing on
the remote server without the user having interacted with it before. They will
need to create a local RSL branch and an initial nonce. However, it remains
unimplemented. It remains unclear whether to give them an initial nonce or to
try the validate repo without a nonce in the local repo. Also, this commit adds
some smarter error handling.