Fix git add and commit nonce bag, write nonce bag
Prior to this commit, the nonce bag implementation was broken. The bag was
mistakenly written to `.git/NONCE_BAG` rather than `./NONCE_BAG`. The path
added to the index was absolute when it should have been relative. The branch
was not checked out correctly. There are still some issues with the index and
work tree not getting updated properly, but at least the nonce bag is now
successfully written and commited during init. However, all other files that
are in the repo but not on the RSL branch are added and committed as well. This
is a problem, as you might imagine.