Create local and remote test directories
Prior to this commit, the test setup only created one test repo. This commit
improves it by creating a mock remote repo and cloning a mock local repo
from it, in order to test functions that involve communication between two
repos. By cloning the local from the remote, the remote is automatically set
as remote `origin`. In addition, the NONCE file is copied directly into the
local developer dir from a separate test ficture, since it shouldn't exist in
the remote repository and doesn't get copied by the clone procedure.