Operation error-chain is a go
After a few fits and starts, some half baked solutions that I turned around
from and then later realized would have worked if only I did something little
and silly, like add a use statement in the errors module (thank you unhelpful
compile errors), we have a working crate that is threaded through with one
great big error chain. Now when things go wrong, we get every error from the
original source on up. Unfortunately, we got rid of a lot of nice custom errors
for Nonce and NonceBag, that could have remained, and instead added them to the
error chain. We can change this again if necessary now that we know how to use
remote_links properly. In addition, we will be able to remove a lot of calls
to `chain_err` now that we have auto-generated conversions for git2 and serde
crates.