checkout RSL after init
Prior to this commit, the RSL branch was checked out in various methods,
including nonce bag write and commit, which need to edit the index. This commit
changes the logic to do one checkout after the RSL has been initialized. The
init routine was moved out of the fetch and push loops, after an initial
attempted fetch of the RSL.