Attempt to address the post-commit index issue
Prior to this commit, there was a problem whereby after the NONCE_BAG file was
added and committed, running a `git status` on the command line showed that the
newly commited file as both staged for deletion and untracked (`git diff`
rightly showing nothing). However, rust2git still showed to repository index
as empty. This issue appears to be a problem for other libgit2 wrappers as well.
https://stackoverflow.com/questions/34599073/git2gos-checkouthead-not-updating-the-index
It is possible that the issue stems from HEAD being updated (making the working
tree dirty) before attempting to update the index or working tree.
This commit contains a number of different attempted solutions for the record,
including the one that seems to actually work, namely performing a git reset to
the commit tree immediately after creating the commit. However, this is not a
perfect solution, as now the index shows as non-empty after the commit+reset. If
the index contains every file no matter what stage, that is fine. but if it
contains only staged files then I am very confused indeed, as when the working
directory shows as clean with no files staged for commit, the index should be
empty.