Commit empty index to tree for RSL init
Prior to this commit, the rsl init routine wrote the existing repository index
to the commit tree. This was a mistake, as it contained all of the files in
the project. This commit corrects the routine to first clear the index,
thereby committing a lovely empty tree and making a genuine orphan commit with
not only no parents but no history as well.