Force checkout
Prior to this commit, checkout left untracked files and weird wonky dual-
tracked untracked files, because we were switching to and from indexes and
work trees with totally different histories. Since the point of this function
is to switch back and forth between RSL and regular project branches, which
have totally different histories, we need to force the move. This is not
something we need to worry about, as any untracked, and/or uncommitted work in
the repository is stashed prior to beginning RSL operations.