Correct implementation of nonce bag read and write
Prior to this commit, the implementation of writing and reafing the nonce bag
from a file was all mixed up. The bag was written as a serialized JSON object
but read as if it were just a bunch of nonce strings separated by newlines.
This issue wasn't caught because there was no unit testing for reading from
and writing to disk. This commit updates read and write to implement the
`NONCE_BAG` file as a long string of unseparated bytes, exactly 32 for each
nonce. They are written to disk in a row without delimiters, and read from the
file 32 bytes at a time until the end of file is reached. Includes unit tests.