Pass in slice rather than vector
Prior to this commit, secure_push took a vector rather than a slice, forcing
the caller to allocate. This commit changes the function to allow an
array of string slices instead. In addition, it adds an integration test
for secure_push, checking that the first real push entry after initialization
is valid by attempting to secure-push a second time.