Redesign rsl initialization
Prior to this commit, the nonce bag was initlialized with a nonce in it.
However, the desired state of the local nonce and nonce bag after init is
unclear from the RSL paper.
In addition, in order for the validate routing to succeed the first time around,
a special case would need to be added to handle the case of there not existing
any previous push entries. This commit addresses both issues by simulating a
prior push entry with the developers initial local nonce in the push entry
nonce bag record, then writing an empty nonce bag before pushing the RSL branch
for the first time. This way, the Nonce bag is empty, the way it should be
immediately following a commit, the developer's local nonce is in the nonce bag
snapshot recorded in the last push entry, and there is a push entry in the log
to serve as the input to the `previous hash` field on the next push entry.