Fix condition in RSL validate routine
Prior to this commit, the RSl validate checked the opposite condition
to decide if the newest entries of the RSL it had just traversed were valid.
As soon as the algorithm reaches an entry that is invalid, it starts folding
`None` into the next iteration of the loop. This means that if we reach the
end of the loop and the value is `None`, the RSL is invalid and the routine
should return false. However, prior to this commit, the routine returned `true`
if the value was `None` and `false` if it contained a hash. This commit
reverses the return condition.