Store head of pushed branch in push entry
Prior to this commit, new push entries were generated using the head of the
current branch (RSL), as a place holder. This commit changes the push entry
generator to store the head of the branch being pushed in the `head` attribute.
This way, when we are fetching branches, we can check for each branch we have
fetched that its tip matches the tip of the last push for that branch according
to the RSL.