Validate and fastforward RSL before adding new fetch entry
Prior to this commit, the `secure-fetch` routine first fetched the
remote RSL, then updated the nonce bag to represent having fetched a new state,
then committed that nonce bag to the local RSL branch, then validated that
RSL, then attempted to push to remote. If there were new push entries, this
order of operations would fail, as the local branch being pushed did not
have the push entries in it. This commit updates the order of operations to
validate and fastforard the rsl prior to updating the nonce bag. The benefits
of this are twofold. While it diverges from the implementation in the RSL
paper, the validation is less trivial. If the validation happens just after you
have put a new nonce in the nonce bag, you are only checking that. However,
if you validate against your old nonce, you are really checking that the last time
somebody pushed, your old nonce was in the bag. THEN you update it. I think this
is what the authors intended, although based on their pseudocode this would
never happen. Secondly, since we validate before editing the RSL, we can
fastforward onto our local branch and do work there, on top of the new entries.
This gets around the failure to push bug without necessitating doing work in
a detached head state.