Stash ignored files
Prior to this commit, the `stash_local_changes` function did not do anything
to ignored files (significantly, builds and compiled binaries and the like).
When RSL checks out a branch with a different history, it also updates the work
tree to match. While this may not be the cleverest thing to do in retrospect,
(for instance it may be preferable to create a new tempdir, build a worktree
from the contents of that dir and commit work from it and then discard it), it
is the thing we do. Consequently, when the original branch is re-checked out,
meaning the branch's commit is read into the index and work tree, we don't get
any of the ignored files back, because surprose surprise, git ignired them!
This commit sticks all those files into the stash with other insaved and
untracked changes, so we can recover them and not rebuild our binaries after
every fetch or push.