Prevent accidental committing of files with a checkout
Prior to this commit, checking out the orphan RSL commit left all of the project
files in the index. All gitignored files were getting deleted when checking out
the original branch once again. In order to alleviate this, the application
stashed ignored files for RSL operations and then popped them at the end. Since
some projects have a lot of ignored files in the form of compiled binaries, this
method was extremely slow. Turns out the problem of deleting files only applied
to the first checkout. This commit changes the init to do some cleanup, performing
a force checkout of the original branch in the middle of init, and coming back to
the now 'born' RSL branch, cleaning up the index without deleting ignored files.