Search for gpghome if not provided
Prior to this commit, if a gpghome dir was not provided as an argument to sign
and verify, we stopped there and let the chips fall where they may. This was
fine for cli sign and verify, for which the cli tool has alternate routes, but
made the gpgme versions nonfunctional. This commit enhances the functions by
having them search elsewhere before giving up. If no homedir is provided, they
first check the GPGHOME environment variable. if that is not set then they use
the gpg default, which is `$HOME/.gnupg`. If no homedir is set, the cli
versions bravely proceed in hopes that the external gpgtools succeed where they
have failed. However, the GPGME functions error, because they will always fail
without a gpg homedir in the context.