Remove libgit2-sys as a direct dependency
At this point we were only using it to get the size of a raw oid byte array,
instead of hardcoding it to 20. In order to integrate the signing functions,
we need to depend on an unstable version of git2rs for the time being, and
can't declare both as dependencies for some reason. When the version of git2rs
is bumped to 0.7.1 we can rewuire that and this problem will go away. However,
we probably want to avaid dealing with libgit2-sys directly as far as possible.
As the number of bytes in an oid is unlikely to change, maybe hard-coding it
is not such a big deal.