Ascii armor gpg signatures
Prior to this commit, the generated signatures were not ascii
armored and therefore not likely to be valid utf8. libgit2 requires
a valid utf8 signature to attach to a commit (and looking at signed commits,
it is obvious that they are ascii). Also, this commit changes the sign and
verify functions to return and take a signature as a string, rather than a
vector. This enables us to pass the resulting signature directly into the
signed commit method, without performing any transformations on the signature
data.