Implement signed commits
Prior to this commit, all commits made on the RSL were unsigned. This commit
implements a wrapper around `git2::Repo::commit` to make signed commits. In
order to do so, the application makes *two* commits. First an unsigned, then
a signed, resetting the target reference of the branch to be updated to point
to the signed commit. This is  less than ideal, as we add a bunch of
garbage to the git object database. We will be able to do this differently when
git2rs implements a wrapper around the libgit2 `create_commit_buffer` method,
which creates a commit without writing it to the object database. While that
will alleviate the need for two commits, we will still have to update the
reference manually (I think).