Refactor RSL struct
Now instead of passign around a big tuple we can do all of our operations
on a nice neat rsl struct which contains info pertinent to the local rsl,
the remote rsl, nrepo nonce and nonce bag, as well as references to the
repo and remote. Since the struct contains mutable references to other
structs, one of which contains a reference to the other, we must be very
explicit about lifetimes, ensuring that the all refs are outlasted by `'repo`
in the end.