Make commit take full refname
Prior to this commit, the commit methods took a short name
and rewuired that name to refer to a local branch. Now that we understand more
clearly what refs can be used where, this commit standardizes the method and
others to take a full reference, so that the heloer functions can be used for
other symbolic refs as well as local branches. This allows us, for example,
to commit dirrectly to a remote ref for testing purposes.