Check that provided refs are already tracked by RSL
Prior to this commit, if a user attempted to fetch a ref that hadn't
previously been secure-pushed, the program would correctly fail. However,
it would leave that branch in `refs/remotes/origin/branch` even though
it is not to be trusted. This commit checks before fetching that the reference
is being tracked by RSL, halting if it is not.