Smarter borrowing
Prior to this commit, we were discovering repos all over the place in the
main function, in order to avoid illegal borrowing. This commit shunts
all of the RSL work into a single code block that _returns_ a value, which
we did not know how to do before. This way, we can do a lot of stuff
with an immutable borrow of a value that is already borrowed mutably, and
save the results of that messing around, and still use the mutable ref
later on. The key is that we are not holding on to any immutable reference
to that value outside of the middle block.