Refactor main function to enable testing
Prior to this commit, there was no way to specify the arguments
and repository which serve as input to RSL operations outside of
the command line parser. This commit separates the parsing and repo discovery
into a different function from performing the operations, so that we may
more easily spoof inputs for testing.