Remove fetch-before-push constraint
Prior to this commit, fetching before push was implemented as a proposed solution to a situation in which metadata has been tampered with and the next interaction is a push -- since pushes do not currently validate reference correctness, a valid push entry could be created attesting to the location of potentially modified references.
This commit removes the fetch portion of push, since fetch will fail if any of the refs being fetched are not present in the remote rsl (which will clearly be the case if it is the first time a reference is being pushed to the remote)
A different solution will need to be added, but for now this commit removes the fetch in order to unblock other work.