Move local, remote dirs into same subdir for test helpers
Prior to this commit, the `setup_fresh` test helper method only supported creating one local repository, since the temp dir itself was the local clone's working directory and the temp dir + `.git` was the remote's directory. This commit changes that to be `{tempdir}/0/.git` and `{tempdir}/central.git` to allow for additional local clones to be stored in the tempdir (i.e. `{tempdir}/1/.git`, `{tempdir}/2/.git` to facilitate testing with multiple local repositories.