Some code cleanup, test refactoring for prettiness
Prior to this commit, there was a lot more duplicated code all over the place.
This commit gets rid of some of that to make the PBT easier to look at.
It also re-adds the git-pbt so that we can pinpoint what in the behavior causes git to detect the metadata tampering.