Serialize prop tests
Cargo tries to run all tests in parallel; this causes the prop tests to fail,
because they try to use the same directory for their git repo. Even if that was
fixed, it's likely that the user-level gpg daemons would not play nicely with
parallel tests.

This commit serializes prop test execution via a mutex.