Add ref specification to git pull
Prior to this commit, git pull would just do a general pull without specifying the branch name. This commit changes it to pull the specific branch requested, enabling repositories without a reference to that branch to pull and create a local tracked version of that branch. This is how the model operates so the functionality needs to be the same. Somehow pulling specific references was removed from the code along the way.