Remove Drop from Context
Prior to this commit, there was a Drop implementation for the Context test helper struct. Since the test helper is made up of temporary directories which clean themselves up when a context goes out of scope, this is unnecessary and could possibly cause problems with how tempdir removes the directories.