Reduce code duplication across binaries
Prior to this commit, code that was commonly used throughout the binaries was not split off into a shared module.
This commit introduces a shared module for the binaries to be able to parse App args into their expected types, discover the current repository, and error handling. This cuts down on some repeated code.